Initial commit of Python script and example output
The Python script I used to create the Proizvolov Emoji cards uses the Python Imaging Module (Pillow). I did not include the emoji PNG files which it needs as input. It is fairly easy to source emoji PNG files -- the script requires these PNG files to have alpha masks. The dimensions and filenames are hardcoded - generalising and customisability is future work

The example output files can be printed using any colour Deskjet printer, cut out, and folded to make the set of 12 cards.